IIC classes and cleanup
Cleaned up the HBridge classes and corrected some errors in the
QuadHBridge.

Added the IIC, Accelerometer, Compass, Gyroscope and Temperature
classes. The default hardware is the unit that I currently have but
additional units can be added as the specifications and testing are
performed.

Also added Example ino's for the IIC library classes.